#212 Move some files around for cohesion
