+++ src/ts/content/extensions/util/tagValidation/index.ts
@@ -1,0 +1,3 @@
+import {appendTagValidator} from "./appendTagValidator";
+
+export {appendTagValidator};
