+++ src/ts/content/extensions/util/tagValidation/types.ts
@@ -1,0 +1,3 @@
+type GetTagsOptions = {noCache: boolean};
+
+export type {GetTagsOptions};
